fix some minor bugs

+++ user/test/test.c
@@ -10,7 +10,7 @@ void kbd_thread(void) {
 int main(int argc, char **argv) {
     printf("subsystem\n");
 
-    universe_syscall(SYS_PIPE_TRIGGER, stdin, &kbd_thread,0,0,0);
+    universe_syscall(SYS_PIPE_TRIGGER, stdin, (uint32_t)&kbd_thread,0,0,0);
 
     while(1){}
 	return 0;
